+++ OpenGL-Engine/src/Application.hpp
@@ -18,9 +18,9 @@ namespace Engine
 		GLFWwindow* m_window = nullptr;						// A reference to the window
 
 		bool m_gladLoaded = false;							// Whether glad has loaded or not
-		unsigned int m_winWidth = 0U, m_winHeight = 0U;		// The width and height of the window
-		unsigned int m_frames = 0U;							// The total amount of frames rendered
-		unsigned int m_fps = 0U, m_framesPerSecond = 0U;	// The amount of frames rendered per second
+		uint16_t m_winWidth = 0U, m_winHeight = 0U;		// The width and height of the window
+		uint64_t m_frames = 0U;							// The total amount of frames rendered
+		uint16_t m_fps = 0U, m_framesPerSecond = 0U;	// The amount of frames rendered per second
 		double m_currentTime = 0.0,
 		 m_prevTime = 0.0, m_deltaTime = 0.0;				// The time between rendered frames
 		double m_fixedTimer = 0.0f, m_frameTimer = 0.0f;	// Timers used for calling fixed update and displaying fps
@@ -45,8 +45,8 @@ namespace Engine
 	public:
 		const double m_fixedDeltaTime = 0.01666666666666666;
 
-		void Run(unsigned int pWidth, unsigned int pHeight, std::string pTitle, bool pFullscreen);
-		void SetDimensions(unsigned int pWidth, unsigned int pHeight);
+		void Run(uint16_t pWidth, uint16_t pHeight, std::string pTitle, bool pFullscreen);
+		void SetDimensions(uint16_t pWidth, uint16_t pHeight);
 		void MouseCallback(double pPosX, double pPosY);
 		void ScrollCallback(double pOffsetX, double pOffsetY);
