+++ OpenGL-Engine/src/Project.cpp
@@ -44,7 +44,7 @@ void Project::Shutdown() {}
 void Project::Update()
 {
 	// Rotates the cubes
-	for (unsigned int i = 0; i < s_numCubes; i++)
+	for (uint8_t i = 0; i < s_numCubes; i++)
 	{
 		float angle = (float)GetDeltaTime() * 5.0f * ((i + 1) / (i * 0.2f + 1));
 		m_cubes[i]->SetTransform(rotate(m_cubes[i]->GetTransform(), radians(angle), glm::normalize(vec3(1.0f, 0.3f, 0.5f))));
